+++ src/components/Workspace.svelte
@@ -34,10 +34,10 @@
 <div class="notebook-container w-11/12 mb-40 md:columns-2 mx-auto mt-5 {currentState}">
 	<div class="h-screen w-max md:w-full inline-block">
 		<div class="relative">
-			<button class="absolute right-2 top-2 z-10" on:click={swapViews}
+			<button class="absolute right-2 top-2 z-10" id="swap" on:click={swapViews}
 				><FontAwesomeIcon
 					icon={faRightLeft}
-					class="text-[#333333] border-[#333333] border-[1px] bg-white dark:bg-stone-200 p-2 rounded-full"
+					class="text-[#333333] border-[#333333] border-[1px] bg-white p-2 rounded-full"
 				/></button
 			>
 		</div>