+++ src/routes/+layout.svelte
@@ -2,12 +2,13 @@
 	import '../app.css';
 	import Header from '../components/Header.svelte';
 	import { darkMode } from '../stores/mode';
-	import { dayTheme } from '../stores/mode';
-
-	document.documentElement.classList.add($dayTheme);
+	import { dayTheme, nightTheme } from '../stores/mode';
 
 	if ($darkMode) {
 		document.documentElement.classList.add($darkMode);
+		document.documentElement.classList.add($nightTheme);
+	} else {
+		document.documentElement.classList.add($dayTheme);
 	}
 </script>